Check that a blog can only be deleted by the owner (4.21). Refactor tests for adding and deleting blogs (4.22)

+++ part4/bloglist/controllers/blogs.js
@@ -45,6 +45,23 @@ blogsRouter.post('/', async (req, res) => {
 });
 
 blogsRouter.delete('/:id', async (req, res) => {
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' });
+  }
+
+  const blog = await Blog.findById(req.params.id);
+  if (!blog) {
+    return res.status(400).json({ error: 'blog not found' });
+  }
+
+  if (blog.user.toString() !== decodedToken.id) {
+    return res.status(401).json({ error: 'invalid user or token' });
+  }
+
+  // NOTE: <document>.remove() seems deprecated
+  // await blog.remove();
   await Blog.findByIdAndRemove(req.params.id);
   res.status(204).end();
 });
